refactor: ScoreCalculator 생성
